add KerberosInterpreter and move kinit loginc there.
